// src/app/shared/models/examination-stat.model.ts
export interface ExaminationStat {
  NGAY_TIEP_NHAN: string;
  THU_TUAN: string;
  TUAN_NAM: number;
  THANG: number;
  NAM: number;
  QUY: number;
  TONG_LUOT_TIEP_NHAN: number;
  BENH_CU: number;
  BENH_MOI: number;
  BHYT: number;
  VIEN_PHI: number;
  LUOT_KHAM_CK: number;
  LUOT_CC: number;
  LUOT_NT: number;
  LUOT_DNT: number;
}